requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.3.0
fastapi==0.118.0
flatbuffers==25.9.23
frozenlist==1.7.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
magic-filter==1.0.12
mpmath==1.3.0
multidict==6.6.4
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
propcache==0.3.2
protobuf==6.33.2
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyreadline3==3.5.4
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==22.5
PyYAML==6.0.3
requests==2.32.5
setuptools==58.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.48.0
style==1.1.0
sympy==1.14.0
typing_extensions==4.12.2
update==0.0.1
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.20.1
onnx>=1.15.0
onnxruntime>=1.21.0
onnxruntime-openvino>=1.21.0; sys_platform != 'darwin'
opencv-python>=4.10
einops>=0.7.0
streamlit>=1.32.0
streamlit-webrtc>=0.47.6
